backend/scrapers/pdf_scraper.py
# backend/scrapers/pdf_scraper.py
"""PDF flyer scraping using OCR."""

from typing import Dict, Any
from .base_scraper import BaseScraper, ScrapingResult, ScrapingMethod

class PDFScraper(BaseScraper):
    """PDF flyer processing using OCR."""
    
    @property
    def method(self) -> ScrapingMethod:
        return ScrapingMethod.PDF_OCR
    
    def scrape_postal_code(self, postal_code: str) -> ScrapingResult:
        """PDF scraping doesn't work with postal codes directly."""
        return ScrapingResult(
            success=False,
            method_used=self.method,
            stores=[],
            offers=[],
            error_message="PDF scraper requires PDF file paths. Use scrape_store with PDF URL."
        )
    
    def scrape_store(self, store_url: str, store_name: str = None) -> ScrapingResult:
        """Scrape PDF flyer using OCR."""
        try:
            # TODO: Implement PDF download and OCR processing
            # - Download PDF from URL
            # - Use pytesseract to extract text
            # - Parse extracted text for product names and prices
            # - Return structured data
            
            return ScrapingResult(
                success=False,
                method_used=self.method,
                stores=[],
                offers=[],
                error_message="PDF OCR scraping not yet implemented"
            )
            
        except Exception as e:
            self.logger.error(f"PDF scraping failed: {str(e)}")
            return ScrapingResult(
                success=False,
                method_used=self.method,
                stores=[],
                offers=[],
                error_message=str(e)
            )
    
    def is_available(self) -> bool:
        """Check if OCR dependencies are available."""
        try:
            import pytesseract
            import PIL
            return True
        except ImportError:
            return False
