backend/requirements.txt
# backend/requirements.txt

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Redis & Job Queue
redis>=5.0.0
celery>=5.3.0

# LangGraph & AI
langgraph>=0.0.30
langchain>=0.1.0
langchain-community>=0.0.10

# LLM Integration (keeping your existing approach)
requests>=2.31.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Telemetry & Monitoring
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-prometheus>=0.54b0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-instrumentation-requests>=0.46b0
prometheus-client>=0.19.0

# Web Scraping & Tools (Selenium-first approach)
selenium>=4.15.0
webdriver-manager>=4.0.0
playwright>=1.40.0  # Keep for advanced cases
lxml>=4.9.0

# File Processing
pypdf2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
pandas>=2.0.0

# Image Processing
pillow>=10.0.0
pytesseract>=0.3.10

# External APIs
googlemaps>=4.10.0
httpx>=0.25.0

# Email notifications
fastapi-mail>=1.4.0

# Utilities
python-slugify>=8.0.0
pytz>=2023.3